chore: fix clippy errors, test failures, and add Makefile for local CI
- Fix usize * float type errors in gromos-core, gromos-io, gromos-integrators,
  gromos-forces benches (cast i as f64 before multiplication)
- Fix Vec3::new(usize, ...) mismatched types in io_bench.rs
- Replace approximate PI/SQRT_2 literals with std::f64::consts constants
- Fix loop-that-never-loops in bin_box.rs and copy_box.rs (-> if let)
- Fix LJParameters struct literal missing cs6/cs12 in nonbonded_bench.rs
- Fix wrong ref_dir() path: GROMOS_references -> gromosXX_references
- Fix test_topology_read_integration: IAC is 0-indexed in memory (6->5)
- Update nacl_pair force tolerance to 1e-3; document deliberate divergence
  from gromosXX due to more precise four_pi_eps_i (138.9354859 vs 138.9354)
- Add Makefile: make check (pre-commit), make ci (full suite = fmt+lint+test)
  — CI was calling make targets that didn't exist (silently failed)

+++ crates/gromos-core/benches/math_bench.rs
@@ -53,7 +53,7 @@ fn bench_nearest_image_scaling(c: &mut Criterion) {
     for n in [100usize, 1_000, 10_000] {
         let pairs: Vec<(Vec3, Vec3)> = (0..n)
             .map(|i| {
-                let t = i * 0.1;
+                let t = i as f64 * 0.1;
                 (
                     Vec3::new(t.sin(), t.cos(), t * 0.01),
                     Vec3::new(t.cos(), t.sin(), t * 0.02),
